git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1028-optimize-loop-detection@42496 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/pointeraliasing/CToFormulaConverterWithPointerAliasing.java
@@ -1350,13 +1350,13 @@ public class CToFormulaConverterWithPointerAliasing extends CtoFormulaConverter 
 
   /** {@inheritDoc} */
   @Override
-  protected int getSizeof(CType pType) {
+  protected long getSizeof(CType pType) {
     return super.getSizeof(pType);
   }
 
   /** {@inheritDoc} */
   @Override
-  protected int getBitSizeof(CType pType) {
+  protected long getBitSizeof(CType pType) {
     return super.getBitSizeof(pType);
   }
